Proof of concept for #1
The reason this is a bit tricky is because we are trying to keep
track of two separate states at once: The "first line/body" state,
and the commited/unstaged changes state.

These can be interleaved in any combination.

General idea is to have separate rule sets for the different "states"

- first-line, "plain text" line is the "summary"
- second-line, "plain text" line is the "second line" (blank)
- body, "plain text" lines are "body" lines

The idea is that while inside any of these states, you can enter a
"changes" section. If you advance the state of the line while inside
one of these sections, you need to bump along to the changes section
for the next line state so that we "remember" which line we are up to
once the changes section is finished.