Reduce duplication via templating
Couldn't get YAML anchors/references to do what I wanted.
Also couldn't figure out how to re-use a `match` but with a different
`set` with pure sublime-syntax constructs. If you "include" a context
you essentially merge its sequence with the current one, but that won't
let you change any of it (e.g. the `set` child of a `match`)

Enter, a quick Custom templating solution that maintains
the current level of indentation and takes a variable for the
`set` context (hardcoded). It allows match/scope/captures to be shared
while bumping along to a different context.